refactor(vault): extract note-reads module
Second step of the vault.ts decomposition (#230, #239). Move the pure note
read operations - readNote, readNoteLineRange (+ NoteLineRangeRead), listNotes -
into a new src/lib/note-reads.ts. Relocate shared traversal primitives
walkVault/filterReadable/isReadAllowed down into vault-fs (both note-reads and
vault.ts need them); move note-domain helper assertMarkdownNotePath into
note-reads, re-imported by vault.ts for the mutations that stay there.

note-reads imports nothing from vault.ts; vault.ts re-exports the moved
symbols so the ./vault.js API and tests are unchanged. typecheck + lint +
970 tests green; no new circular dependency.
